wait for events in the view
